split off node request file again, added descriptive message when node request access is turned off and added test case
